starting to repair and simplify testing
