+++ case_studies/strong_lensing/config/config.yaml
@@ -32,8 +32,7 @@ datasets:
             run: 94
             camcol: 1
             field: 12
-            bands:
-                - 2
+            bands: [2]
         n_batches: 10
         batch_size: 32
         generate_device: "cuda:0"
@@ -73,14 +72,13 @@ models:
         border_padding: 24
         psf_params_file: ${paths.data}/sdss/5313/3/117/psField-005313-3-0117.fits
         galaxy_model:
-            _target_: bliss.models.galsim_decoder.SingleGalsimGalaxyDecoder
+            _target_: bliss.models.galsim_decoder.GalsimGalaxyDecoder
             n_bands: 1
             slen: 53
             pixel_scale: 0.396
             psf_params_file: ${paths.data}/sdss/5313/3/117/psField-005313-3-0117.fits
             psf_slen: 25
-            sdss_bands:
-                - 2
+            sdss_bands: [2]
         lens_model:
             _target_: bliss.models.galsim_decoder.SingleLensedGalsimGalaxyDecoder
             n_bands: 1
@@ -88,11 +86,9 @@ models:
             pixel_scale: 0.396
             psf_params_file: ${paths.data}/sdss/5313/3/117/psField-005313-3-0117.fits
             psf_slen: 25
-            sdss_bands:
-                - 2
+            sdss_bands: [2]
         psf_slen: 25
-        sdss_bands:
-            - 2
+        sdss_bands: [2]
     detection_encoder:
         _target_: bliss.models.detection_encoder.DetectionEncoder
         input_transform:
@@ -131,16 +127,8 @@ models:
         dropout: 0.0
         optimizer_params:
             lr: 1e-4
-    galaxy_net:
-        _target_: bliss.models.galaxy_net.OneCenteredGalaxyAE
-        slen: 53
-        latent_dim: 8
-        hidden: 256
-        n_bands: 1
-        optimizer_params:
-            lr: 1e-4
     prior:
-        _target_: bliss.models.prior.ImagePrior
+        _target_: bliss.simulator.ImagePrior
         n_bands: 1
         max_sources: 1
         mean_sources: 0.03
@@ -151,23 +139,19 @@ models:
         prob_galaxy: 0.7
         prob_lensed_galaxy: 0.2
         galaxy_prior:
-            _target_: bliss.models.galsim_decoder.SingleGalsimGalaxyPrior
-            flux_sample: "pareto"
+            _target_: bliss.simulator.prior.GalsimGalaxyPrior
             min_flux: 622.0
             max_flux: 1e6
             alpha: 0.47
-            a_sample: "gamma"
             a_concentration: 0.39330758068481686
             a_loc: 0.8371888967872619
             a_scale: 4.432725319432478
             a_bulge_disk_ratio: 2.0
         lensed_galaxy_prior:
-            _target_: bliss.models.galsim_decoder.SingleGalsimGalaxyPrior
-            flux_sample: "pareto"
+            _target_: bliss.simulator.prior.GalsimGalaxyPrior
             min_flux: 622.0
             max_flux: 1e4
             alpha: 0.47
-            a_sample: "gamma"
             a_concentration: 0.39330758068481686
             a_loc: 0.8371888967872619
             a_scale: 4.432725319432478
