+++ case_studies/summer_template/config.yaml
@@ -133,13 +133,13 @@ encoder:
 training:
     name: "sdss_encoder"
     version: null
-    n_epochs: 1000
+    n_epochs: 1
     experiment: default
     save_top_k: 1
     trainer:
         _target_: pytorch_lightning.Trainer
         logger: True
-        enable_checkpointing: False
+        enable_checkpointing: True
         profiler: null
         reload_dataloaders_every_n_epochs: 0
         max_epochs: ${training.n_epochs}
@@ -148,7 +148,7 @@ training:
         devices: 1
         limit_train_batches: 1.0
         limit_val_batches: 1.0
-        check_val_every_n_epoch: 10
+        check_val_every_n_epoch: 1
         log_every_n_steps: 10  # corresponds to n_batches
     testing:
         file: null
@@ -157,7 +157,7 @@ training:
     weight_save_path: ${paths.project}/models/${training.name}.pt
     seed: 42
 
-inference:
+predict:
     dataset:
         _target_: bliss.surveys.sdss.SloanDigitalSkySurvey
         sdss_dir: ${paths.sdss}
