+++ tests/star_basic/star_basic.yaml
@@ -17,104 +17,87 @@ paths:
     sdss: ${paths.root}/data/sdss
     data: ${paths.root}/data
 
-datasets:
-    simulated_single_tile:
-        _target_: bliss.datasets.simulated.SimulatedDataset
-        prior: ${models.prior}
-        decoder: ${models.decoder}
-        n_tiles_h: 1
-        n_tiles_w: 1
-        n_batches: 10
-        batch_size: 640
-        generate_device: "cuda:0"
-        background:
-            _target_: bliss.datasets.simulated.ConstantBackground
-            background:
-                - 865.0
-
-    simulated:
-        _target_: bliss.datasets.simulated.SimulatedDataset
-        prior: ${models.prior}
-        decoder: ${models.decoder}
-        n_tiles_h: 5
-        n_tiles_w: 5
-        n_batches: 10
-        batch_size: 32
-        generate_device: "cuda:0"
-        background:
-            _target_: bliss.datasets.simulated.ConstantBackground
-            background:
-                - 865.0
-
-decoders:
-    single_tile:
-        _target_: bliss.models.decoder.ImageDecoder
+simulator:
+    _target_: bliss.simulator.simulated_dataset.SimulatedDataset
+    prior:
+        _target_: bliss.simulator.prior.ImagePrior
         n_bands: 1
-        tile_slen: 10
-        ptile_slen: 10
-        border_padding: 0
-        psf_params_file: ${paths.sdss}/94/1/12/psField-000094-1-0012.fits
-        psf_slen: 25
-        sdss_bands:
-            - 2
-
-    star_basic:
-        _target_: bliss.models.decoder.ImageDecoder
-        n_bands: 1
-        tile_slen: 2
-        ptile_slen: 6
-        border_padding: 2
-        psf_slen: 25
-        psf_params_file: ${paths.sdss}/94/1/12/psField-000094-1-0012.fits
-        sdss_bands:
-            - 2
-priors:
-    star_single_tile:
-        _target_: bliss.models.prior.ImagePrior
-        n_bands: 1
-        prob_galaxy: 0.0
-        max_sources: 2
-        mean_sources: 1.2
-        min_sources: 0
-        f_min: 1e5
-        f_max: 1e6
-        alpha: 0.5
-        galaxy_prior: null
-
-    star_basic:
-        _target_: bliss.models.prior.ImagePrior
-        n_bands: 1
-        prob_galaxy: 0.0
         max_sources: 1
         mean_sources: 0.015
         min_sources: 0
         f_min: 1e5
         f_max: 1e6
         alpha: 0.5
+        prob_galaxy: 0.0
         galaxy_prior: null
+    decoder:
+        _target_: bliss.simulator.decoder.ImageDecoder
+        n_bands: 1
+        tile_slen: 2
+        ptile_slen: 6
+        border_padding: 2
+        psf_slen: 25
+        sdss_bands: [2]
+        psf_params_file: ${paths.data}/sdss/94/1/12/psField-000094-1-0012.fits
+    background:
+        _target_: bliss.simulator.background.ConstantBackground
+        background: [865.0]
+    n_tiles_h: 5
+    n_tiles_w: 5
+    n_batches: 10
+    batch_size: 32
+    fix_validation_set: true
+    valid_n_batches: 10  # 256
+    generate_device: "cpu"
+    num_workers: 0  # can only use > 0 workers if simulator is running on the cpu
 
-models:
-    prior: ${priors.star_basic}
-    decoder: ${decoders.star_basic}
-    detection_encoder:
-        _target_: bliss.models.detection_encoder.DetectionEncoder
-        input_transform:
-            _target_: bliss.models.detection_encoder.LogBackgroundTransform
-        n_bands: ${models.decoder.n_bands}
-        tile_slen: ${models.decoder.tile_slen}
-        ptile_slen: ${models.decoder.ptile_slen}
-        max_detections: ${models.prior.max_sources}
-        channel: 8
-        spatial_dropout: 0.0
-        dropout: 0.0
-        hidden: 128
-        annotate_probs: false
-        optimizer_params:
-            lr: 1e-3
+encoder:
+    _target_: bliss.encoder.Encoder
+    n_bands: ${simulator.decoder.n_bands}
+    tile_slen: ${simulator.decoder.tile_slen}
+    tiles_to_crop: 0
+    annotate_probs: True
+    slack: 1.0
+    optimizer_params:
+        lr: 1e-3
+    scheduler_params:
+        milestones: [32]
+        gamma: 0.1
+    architecture:
+        # this architecture is based on yolov5l.yaml, see
+        # https://github.com/ultralytics/yolov5/blob/master/models/yolov5l.yaml
+        depth_multiple: 1.0  # model depth multiple
+        width_multiple: 1.0  # layer channel multiple
+        anchors:
+            # TODO: investigate whether other size anchors would work better
+            - [4, 4]  # P3/8
+        backbone: [
+            # [from, number, module, args]
+            [-1, 1, Conv, [64, 5, 1]],
+            [-1, 3, Conv, [64, 1, 1]],
+            [-1, 1, Conv, [128, 3, 2]],
+            [-1, 1, Conv, [128, 3, 1]],
+            [-1, 1, Conv, [256, 3, 2]],
+            [-1, 6, C3, [256]],
+            [-1, 1, Conv, [512, 3, 2]],
+            [-1, 9, C3, [512]],
+            [-1, 1, Conv, [1024, 3, 2]],
+            [-1, 3, C3, [1024]],
+            [-1, 1, SPPF, [1024, 5]],
+        ]
+        head: [
+            [-1, 1, Conv, [512, 1, 1]],
+            [-1, 1, nn.Upsample, [None, 2, 'nearest']],
+            [[-1, 6], 1, Concat, [1]],
+            [-1, 3, C3, [512, false]],
+            [-1, 1, Conv, [256, 1, 1]],
+            [-1, 1, nn.Upsample, [None, 2, 'nearest']],
+            [[-1, 4, 5], 1, Concat, [1]],
+            [-1, 3, C3, [256, false]],
+            [[17], 1, Detect, [nc, anchors]],
+        ]
 
 training:
-    model: "${models.detection_encoder}"
-    dataset: "${datasets.simulated}"
     n_epochs: 201
     experiment: default
     version: null
@@ -125,7 +108,7 @@ training:
         enable_checkpointing: False
         profiler: null
         reload_dataloaders_every_n_epochs: 0
-        check_val_every_n_epoch: 50
+        check_val_every_n_epoch: 1001
         max_epochs: ${training.n_epochs}
         min_epochs: ${training.n_epochs}
         accelerator: "gpu"
@@ -136,4 +119,4 @@ training:
         batch_size: 32
         num_workers: 0
     weight_save_path: null
-    seed: null
+    seed: 42
