+++ tests/star_basic/test_sleep_star.py
@@ -4,16 +4,13 @@ from bliss.train import train
 
 
 # TODO: add some tests of blended stars and galaxies
-class TestSleepStarOneTile:
+# TODO: add a test with one tile
+class TestStarBasic:
     @pytest.fixture(scope="class")
     def overrides(self, devices):
-        generate_device = "cuda:0" if devices.use_cuda else "cpu"
         return {
-            "datasets.simulated_single_tile.generate_device": generate_device,
-            "models.prior": "${priors.star_single_tile}",
-            "models.decoder": "${decoders.single_tile}",
-            "training.dataset": "${datasets.simulated_single_tile}",
             "training.n_epochs": 201 if devices.use_cuda else 2,
+            "simulator.generate_device": "cuda:0" if devices.use_cuda else "cpu",
         }
 
     @pytest.fixture(scope="class")
@@ -24,7 +21,7 @@ class TestSleepStarOneTile:
         results = star_basic_model_setup.test_model(overrides, trained_sleep)
         assert {"precision", "f1", "avg_distance"}.issubset(results.keys())
 
-        # only expect tests to pass if gpu
+        # we only expect the tests below to pass if we're training on the gpu
         if not devices.use_cuda:
             return
 
@@ -32,31 +29,5 @@ class TestSleepStarOneTile:
         assert results["f1"] > 0.8
         assert results["avg_distance"] < 1.0
 
-
-class TestSleepStarTiles:
-    @pytest.fixture(scope="class")
-    def overrides(self, devices):
-        return {
-            "training.n_epochs": 201 if devices.use_cuda else 2,
-            "datasets.simulated.generate_device": "cuda:0" if devices.use_cuda else "cpu",
-        }
-
-    @pytest.fixture(scope="class")
-    def trained_sleep(self, overrides, star_basic_model_setup):
-        return star_basic_model_setup.get_trained_model(overrides)
-
-    def test_simulated(self, overrides, trained_sleep, star_basic_model_setup, devices):
-        results = star_basic_model_setup.test_model(overrides, trained_sleep)
-        assert {"precision", "f1", "avg_distance"}.issubset(results.keys())
-
-        # only expect tests to pass if gpu
-        if not devices.use_cuda:
-            return
-
-        assert results["precision"] > 0.85
-        assert results["f1"] > 0.8
-        assert results["avg_distance"] < 1.0
-
-    def test_train(self, overrides, star_basic_model_setup):
-        cfg = star_basic_model_setup.get_cfg(overrides)
+    def test_train(self, cfg):
         train(cfg)
